pyproject.toml
[project]
name = "research-weekly-feed"
version = "0.1.0"
description = "Research paper aggregator for computational methods, gender inequality, and social media"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.0.0",
    "feedparser>=6.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
